requirements.txt
typer>=0.9.0
python-dotenv>=1.0.0
docker>=6.0.0
python-telegram-bot>=20.0
discord.py>=2.3.0
openai>=1.0.0
chromadb>=0.4.0
apscheduler>=3.10.0
pyyaml>=6.0
python-dateutil>=2.8.0
ddgs>=7.0.0
httpx>=0.24.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
tiktoken>=0.5.0
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=12.0
